--- app/src/services/api.ts
@@ -1,2 +1,0 @@
-export { api } from '../../../renderer/src/services/api'
-
