--- app/src/types/index.ts
@@ -1,2 +1,0 @@
-export * from '../../../renderer/src/types'
-
