+++ renderer/src/components/SugestoesPage.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useMemo, useState } from 'react';
-import { api } from '../../../app/src/services/api';
-import type { Designacao, Saida, Territorio } from '../../../app/src/types';
+import { api } from '../services/api';
+import type { Designacao, Saida, Territorio } from '../../../models';
 import SugestoesTable, { Sugestao } from './SugestoesTable';
 
 const MESES_LIMITE = 5;
